requirements.txt
# Core Orchestration
langgraph>=0.2.0
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0

# Data & Database
pandas>=2.2.0
sqlalchemy>=2.0.0
tabulate>=0.9.0  # Useful for debugging DF output

# Visualization & UI
streamlit>=1.36.0
plotly>=5.22.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
tiktoken>=0.7.0  # For token estimation if needed